backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pydantic
httpx
stripe
firebase-admin
python-dotenv
Pillow
openai
langchain
pytesseract
whisper 
python-dotenv